`Enhancement:` moving logic to hooks
