+++ src/components/layouts/LayoutStandard/index.tsx
@@ -1,22 +1,17 @@
-import { CSSProperties, ReactElement, useEffect } from 'react'
-
 import Footer from '@blocks/Footer'
 import Header from '@blocks/Header'
 import Main from '@components/ui/Main'
+import { ReactElement } from 'react'
 import Trigger from '@components/ui/Trigger'
-import { getScrollBarWidth } from '@utilities/scrollBarWidth'
 import styles from './Site.module.css'
+import { useScrollBarWidth } from '@hooks/useScrollBarWidth'
 
 type props = {
   children: ReactElement
 }
 
-const area = (value: string) => ({ ['--local-grid-area']: `${value}` })
-
 export default function LayoutStandard({ children }: props) {
-  useEffect(() => {
-    getScrollBarWidth()
-  }, [])
+  useScrollBarWidth()
 
   return (
     <div className={styles.site}>
