+++ src/hooks/useScrollToTop.ts
@@ -1,0 +1,9 @@
+import { useEffect } from "react"
+
+export const useScrollToTop = () => {
+  useEffect(() => {
+    window.scrollTo(0, 0)
+
+    return () => window.scrollTo(0, 0)
+  }, [])
+}
