+++ src/pages/account/index.tsx
@@ -1,8 +1,11 @@
 import LayoutStandard from '@layouts/LayoutStandard'
 import Meta from '@head/Meta'
 import { NextPage } from 'next'
+import { useScrollToTop } from '@hooks/useScrollToTop'
 
 const Account: NextPage = () => {
+  useScrollToTop()
+  
   return (
     <LayoutStandard>
       <>
