+++ src/pages/account/profile.tsx
@@ -1,8 +1,11 @@
 import LayoutStandard from '@layouts/LayoutStandard'
 import Meta from '@head/Meta'
 import { NextPage } from 'next'
+import { useScrollToTop } from '@hooks/useScrollToTop'
 
 const Profile: NextPage = () => {
+  useScrollToTop()
+  
   return (
     <LayoutStandard>
       <>
