--- src/utilities/scrollBarWidth.ts
@@ -1,6 +1,0 @@
-export const getScrollBarWidth = () => {
-  let root = document.documentElement
-  const width = window.innerWidth - root.clientWidth
-
-  root.style.setProperty('--scrollbar-width', `${width}px`)
-}
